+++ tests/test_vae.py
@@ -3,25 +3,10 @@ import time
 
 import jax.numpy as jnp
 import matplotlib.pyplot as plt
-import numpy as np
-from sqlalchemy.sql.coercions import expect
 
 from starccato_jax import Config, StarccatoVAE
-from starccato_jax.data import get_default_weights, load_training_data
-from starccato_jax.plotting import plot_distributions
+from starccato_jax.data import get_default_weights
 from starccato_jax.vae.core.io import TrainValMetrics, load_loss_h5
-
-
-def test_version():
-    from starccato_jax import __version__
-
-    assert isinstance(__version__, str)
-
-
-def test_training_data():
-    train_data, val_data = load_training_data(train_fraction=0.8, clean=True)
-    assert train_data.shape == (1411, 256)
-    assert isinstance(train_data, jnp.ndarray)
 
 
 def test_train_vae(outdir):
@@ -29,7 +14,6 @@ def test_train_vae(outdir):
     t0 = time.time()
     vae = StarccatoVAE.train(
         model_dir=outdir,
-        train_fraction=0.8,
         config=Config(latent_dim=8, epochs=10, cyclical_annealing_cycles=0),
     )
     runtime = round(time.time() - t0, 2)