+++ tests/test_version.py
@@ -1,0 +1,5 @@
+def test_version():
+    from starccato_jax import __version__, __version_tuple__
+
+    assert isinstance(__version__, str)
+    assert isinstance(__version_tuple__, tuple)